レポート/api_source/requirements.txt
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
requests
flask
flask-cors
gunicorn
firebase-admin
